requirements.txt
pytest==8.1.1
Appium-Python-Client==4.0.0

decouple~=0.0.7
selenium~=4.18.1
allure-pytest~=2.13.2